+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/service/MicrophoneService.kt
@@ -1,8 +1,3 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
 package com.amazonaws.services.chime.sdkdemo.service
 
 import android.app.NotificationChannel
@@ -17,22 +12,20 @@ import android.os.IBinder
 import androidx.core.app.NotificationCompat
 import com.amazonaws.services.chime.sdkdemo.R
 
-class ScreenCaptureService : Service() {
+class MicrophoneService : Service() {
     private lateinit var notificationManager: NotificationManager
+    private val CHANNEL_ID = "MicrophoneServiceChannel"
+    private val CHANNEL_NAME = "Microphone Service Channel"
+    private val NOTIFICATION_ID = 1
 
-    private val CHANNEL_ID = "ScreenCaptureServiceChannelID"
-    private val CHANNEL_NAME = "Screen Share"
-    private val SERVICE_ID = 1
+    private val binder = MicrophoneBinder()
 
-    private val binder = ScreenCaptureBinder()
-
-    inner class ScreenCaptureBinder : Binder() {
-        fun getService(): ScreenCaptureService = this@ScreenCaptureService
+    inner class MicrophoneBinder : Binder() {
+        fun getService(): MicrophoneService = this@MicrophoneService
     }
 
     override fun onCreate() {
         super.onCreate()
-
         notificationManager =
             applicationContext.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
     }
@@ -47,30 +40,31 @@ class ScreenCaptureService : Service() {
             notificationManager.createNotificationChannel(channel)
         }
 
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
             startForeground(
-                SERVICE_ID,
+                NOTIFICATION_ID,
                 NotificationCompat.Builder(this, CHANNEL_ID)
                     .setSmallIcon(R.mipmap.ic_launcher)
-                    .setContentTitle(getString(R.string.screen_capture_notification_tile))
-                    .setContentText(getText(R.string.screen_capture_notification_text))
+                    .setContentTitle(getString(R.string.microphone_notification_tile))
+                    .setContentText(getText(R.string.microphone_notification_text))
                     .setPriority(NotificationCompat.PRIORITY_DEFAULT)
                     .build(),
-                ServiceInfo.FOREGROUND_SERVICE_TYPE_MEDIA_PROJECTION
+                ServiceInfo.FOREGROUND_SERVICE_TYPE_MICROPHONE
             )
         } else {
             startForeground(
-                SERVICE_ID,
+                NOTIFICATION_ID,
                 NotificationCompat.Builder(this, CHANNEL_ID)
                     .setSmallIcon(R.mipmap.ic_launcher)
-                    .setContentTitle(getString(R.string.screen_capture_notification_tile))
-                    .setContentText(getText(R.string.screen_capture_notification_text))
+                    .setContentTitle(getString(R.string.microphone_notification_tile))
+                    .setContentText(getText(R.string.microphone_notification_text))
                     .setPriority(NotificationCompat.PRIORITY_DEFAULT)
                     .build()
             )
         }
+
         return START_STICKY
     }
 
-    override fun onBind(intent: Intent?): IBinder? = binder
+    override fun onBind(intent: Intent?): IBinder = binder
 }